// src/app/modules/contractor/assignment-details/constants/enum.ts
export enum OrganizationLevelEnum {
  One = 'OrganizationLevel1',
  Two = 'OrganizationLevel2',
  Three = 'OrganizationLevel3',
  Four = 'OrganizationLevel4'
}

export enum OrganizationLevelValueEnum{
  One = 'Organization Level 1',
  Two = 'Organization Level 2',
  Three = 'Organization Level 3',
  Four = 'Organization Level 4'
}

export enum ApproverEnum{
  One = 'Approver1',
  Two = 'Approver2'
}

export enum ApproverValueEnum{
  One = 'Approver 1',
  Two = 'Approver 2'
}

export enum RevisionStatus{
  One = 'decline',
  Two = 'withdraw',
  Three = 'approved',
  Four = 'processed'
}